service refactored to handle authorization exceptions in asynchronous threads (combination of @PreAuthorize and @Async does not propagate access exceptions to calling thread)
