+++ wls-ergebnismeldung-service/src/test/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/awerte/AWerteServiceSecurityTest.java
@@ -142,15 +142,10 @@ class AWerteServiceSecurityTest {
         }
 
         @Test
-        void should_failWithAccessDeniedException_when_adminLoadwahltermindatenAuthorityIsMissing() throws Exception {
+        void should_failWithAccessDeniedException_when_adminLoadwahltermindatenAuthorityIsMissing() {
             SecurityUtils.runWith();
             val wahlbezirkID = "wahlbezirkID";
             val wahlbezirkIDList = List.of(wahlbezirkID);
-            val eaiWahlberechtigte = getAWerteForWahlbezirkID(wahlbezirkID);
-
-            WireMock.stubFor(WireMock.get("/wahldaten/wahlbezirke/" + wahlbezirkID + "/wahlberechtigte")
-                    .willReturn(WireMock.aResponse().withHeader("Content-Type", "application/json").withStatus(HttpStatus.OK.value())
-                            .withBody(objectMapper.writeValueAsBytes(eaiWahlberechtigte))));
 
             Assertions.assertThatThrownBy(() -> aWerteService.initialiseAWerte(wahlbezirkIDList)).isInstanceOf(AccessDeniedException.class);
         }
