+++ wls-ergebnismeldung-service/src/test/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/awerte/AWerteServiceTest.java
@@ -2,7 +2,6 @@ package de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.service.awerte;
 
 import static org.mockito.Mockito.times;
 
-import ch.qos.logback.classic.Level;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.domain.awerte.AWerte;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.domain.awerte.AWerteRepository;
 import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.exception.ExceptionConstants;
@@ -10,7 +9,6 @@ import de.muenchen.oss.wahllokalsystem.ergebnismeldungservice.utils.LoggerExtens
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.TechnischeWlsException;
 import de.muenchen.oss.wahllokalsystem.wls.common.exception.util.ExceptionFactory;
 import de.muenchen.oss.wahllokalsystem.wls.common.security.domain.BezirkUndWahlID;
-import java.util.Base64;
 import java.util.List;
 import lombok.val;
 import org.assertj.core.api.Assertions;
@@ -25,9 +23,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class AWerteServiceTest {
-
-    @Mock
-    AsyncProgress asyncProgress;
 
     @Mock
     AWerteRepository aWerteRepository;
@@ -62,7 +57,9 @@ class AWerteServiceTest {
 
             val result = unitUnderTest.getAWerte(wahlbezirkID);
 
-            Assertions.assertThatCode(() -> unitUnderTest.getAWerte(wahlbezirkID)).doesNotThrowAnyException();
+            Mockito.verify(aWerteValidator, times(1)).validWahlbezirkIDParamOrThrow(Mockito.any());
+            Mockito.verify(aWerteModelMapper, times(1)).fromListOfAWerteModeltoListOfAWerteEntity(Mockito.any());
+
             Assertions.assertThat(result).isEqualTo(aWerteModelListFromClient);
         }
 
@@ -109,63 +106,6 @@ class AWerteServiceTest {
                     .thenReturn(mockedWlsException);
 
             Assertions.assertThatThrownBy(() -> unitUnderTest.getAWerte(wahlbezirkID)).isSameAs(mockedWlsException);
-        }
-    }
-
-    @Nested
-    class InitialiseAWerte {
-
-        @Test
-        void should_notFailAndSaveInDB_when_clientDataFound() {
-            val wahlbezirkID1 = "wahlbezirkID1";
-            val aWerteModelListFromClient1 = createListOfAWerteModels(wahlbezirkID1);
-            val wahlbezirkID2 = "wahlbezirkID2";
-            val aWerteModelListFromClient2 = createListOfAWerteModels(wahlbezirkID2);
-            val wahlbezirkID3 = "wahlbezirkID3";
-            val aWerteModelListFromClient3 = createListOfAWerteModels(wahlbezirkID3);
-            val wahlbezirkIDs = List.of(wahlbezirkID1, wahlbezirkID2, wahlbezirkID3);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID1)).thenReturn(aWerteModelListFromClient1);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID2)).thenReturn(aWerteModelListFromClient2);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID3)).thenReturn(aWerteModelListFromClient3);
-
-            Assertions.assertThatNoException().isThrownBy(() -> unitUnderTest.initialiseAWerte(wahlbezirkIDs));
-            Mockito.verify(aWerteRepository, times(3)).saveAll(Mockito.any());
-        }
-
-        @Test
-        void should_notFailAndSaveInDB_when_base64ClientDataFound() {
-            val wahlbezirkID1 = Base64.getEncoder().encodeToString("WAHLBEZIRK-1410\",\"wahlterminId".getBytes());
-            val aWerteModelListFromClient1 = createListOfAWerteModels(wahlbezirkID1);
-            val wahlbezirkID2 = Base64.getEncoder().encodeToString("WAHLBEZIRK-1411\",\"wahlterminId".getBytes());
-            val aWerteModelListFromClient2 = createListOfAWerteModels(wahlbezirkID2);
-            val wahlbezirkID3 = Base64.getEncoder().encodeToString("WAHLBEZIRK-1412\",\"wahlterminId".getBytes());
-            val aWerteModelListFromClient3 = createListOfAWerteModels(wahlbezirkID3);
-            val wahlbezirkIDs = List.of(wahlbezirkID1, wahlbezirkID2, wahlbezirkID3);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID1)).thenReturn(aWerteModelListFromClient1);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID2)).thenReturn(aWerteModelListFromClient2);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID3)).thenReturn(aWerteModelListFromClient3);
-
-            Assertions.assertThatNoException().isThrownBy(() -> unitUnderTest.initialiseAWerte(wahlbezirkIDs));
-            Mockito.verify(aWerteRepository, times(3)).saveAll(Mockito.any());
-        }
-
-        @Test
-        void should_notFailButLogError_when_clientDataFoundAndRepoFails() {
-            val wahlbezirkID1 = "wahlbezirkID1";
-            val aWerteModelListFromClient1 = createListOfAWerteModels(wahlbezirkID1);
-            val wahlbezirkID2 = "wahlbezirkID2";
-            val aWerteModelListFromClient2 = createListOfAWerteModels(wahlbezirkID2);
-            val wahlbezirkID3 = "wahlbezirkID3";
-            val aWerteModelListFromClient3 = createListOfAWerteModels(wahlbezirkID3);
-            val wahlbezirkIDs = List.of(wahlbezirkID1, wahlbezirkID2, wahlbezirkID3);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID1)).thenReturn(aWerteModelListFromClient1);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID2)).thenReturn(aWerteModelListFromClient2);
-            Mockito.when(aWerteClient.getAWerte(wahlbezirkID3)).thenReturn(aWerteModelListFromClient3);
-            Mockito.when(aWerteRepository.saveAll(Mockito.any())).thenThrow(new RuntimeException());
-
-            Assertions.assertThatNoException().isThrownBy(() -> unitUnderTest.initialiseAWerte(wahlbezirkIDs));
-            Mockito.verify(aWerteRepository, times(3)).saveAll(Mockito.any());
-            Assertions.assertThat(loggerExtension.getLoggedEventsStream().filter(event -> event.getLevel() == Level.ERROR).count()).isEqualTo(3);
         }
     }
